now with a lot less sucking.

+++ tribble/purveyors/zone_status.py
@@ -8,8 +8,13 @@ class ZoneState(object):
         self.cell = cell
         self.schematic = db.get_schematic_id(sid=self.cell['schematic_id'],
                                         uid=self.cell['auth_id'])
-        self.zone = db.get_zones_by_id(skm=self.schematic,
-                                       zid=self.cell['zone_id'])
+        if self.cell.get('zone_id'):
+            self.zone = db.get_zones_by_id(skm=self.schematic,
+                                           zid=self.cell['zone_id'])
+
+    def _reconfig(self):
+        self.cell['zone_state'] = 'RECONFIGURING'
+        self.state_update()
 
     def _build(self):
         self.cell['zone_state'] = 'BUILDING'
@@ -27,14 +32,25 @@ class ZoneState(object):
         self.cell['zone_state'] = 'DELETING'
         self.state_update()
 
-    def _schematic_delete(self):
-        key = db.get_instanceskeys(zon=self.zone)
-        _con = db.get_configmanager(skm=self.schematic)
+    def _delete_resource(self, skm=False):
         sess = _DB.session
-        sess = db.delete_item(session=sess, item=self.zone)
-        sess = db.delete_item(session=sess, item=key)
-        sess = db.delete_item(session=sess, item=self.schematic)
-        sess = db.delete_item(session=sess, item=_con)
+        try:
+            ints = db.get_instances(zon=self.zone)
+            if not len(ints) == 0:
+                self.cell['zone_state'] = 'DELETE FAILED'
+                self.cell['zone_msg'] = ('Found Instance when they should'
+                                         ' have all been deleted')
+                self.state_update()
+            sess = db.delete_item(session=sess, item=self.zone)
+            key = db.get_instanceskeys(zon=self.zone)
+            sess = db.delete_item(session=sess, item=key)
+        except AttributeError, exp:
+            LOG.info('No Zone To Delete as No Zone was Found ==> %s' % exp)
+
+        if skm:
+            _con = db.get_configmanager(skm=self.schematic)
+            sess = db.delete_item(session=sess, item=self.schematic)
+            sess = db.delete_item(session=sess, item=_con)
         db.commit_session(session=sess)
 
     def state_update(self):
